requirements.txt
pandas==1.3.0
numpy==1.19.5
scikit-learn==1.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
seaborn==0.11.1
cbor2==5.4.2
tensorflow==2.6.0
tensorboard==2.6.0
streamlit==1.2.0
edge_impulse_linux==1.0.6
tqdm==4.62.3
pyserial==3.5
